refactor(allocator): flatten primary_allocator & metadata_allocator into allocator module;
Promotes all primary_allocator submodules to direct children of allocator, collapses metadata_allocator/ directory into a single file by inlining region.rs, and rewrites import paths to use `super::` instead of absolute crate paths.
